+++ check_all.sh
@@ -5,9 +5,9 @@ black . --check
 echo "mypy check"
 mypy tensorcircuit
 echo "pylint check"
-pylint tensorcircuit tests
+pylint tensorcircuit tests examples/*.py
 echo "pytest check"
 pytest -n 4 --cov=tensorcircuit -vv -W ignore::DeprecationWarning
 echo "sphinx check"
-cd docs && sphinx-build source build/html
-echo "all checks passed, congratulates! 💐"
+cd docs && sphinx-build source build/html && sphinx-build source -D language="cn" build/html_cn
+echo "all checks passed, congratulation! 💐"