+++ examples/adiabatic_vqnhe.py
@@ -1,3 +1,6 @@
+"""
+Calculate molecule dissociation curve using VQNHE
+"""
 from functools import partial
 import sys